requirements.txt
streamlit
feedparser
google-generativeai
PyGithub
pandas
python-dotenv
